+++ src/main/java/searchengine/model/entities/SiteEntity.java
@@ -1,4 +1,4 @@
-package searchengine.model;
+package searchengine.model.entities;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
