+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import searchengine.model.PageEntity;
-import searchengine.model.SiteEntity;
+import searchengine.model.entities.PageEntity;
+import searchengine.model.entities.SiteEntity;
 
 import java.util.List;
 import java.util.Optional;
