+++ src/main/java/searchengine/repositories/SiteRepository.java
@@ -2,7 +2,7 @@ package searchengine.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import searchengine.model.SiteEntity;
+import searchengine.model.entities.SiteEntity;
 
 import java.util.List;
 import java.util.Optional;
