+++ src/main/java/searchengine/services/PageSearcherService.java
@@ -3,8 +3,9 @@ package searchengine.services;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import searchengine.model.LemmaEntity;
-import searchengine.model.SiteEntity;
+import searchengine.model.Page;
+import searchengine.model.entities.LemmaEntity;
+import searchengine.model.entities.SiteEntity;
 import searchengine.repositories.IndexRepository;
 import searchengine.repositories.LemmaRepository;
 import searchengine.repositories.PageRepository;
