+++ src/main/java/searchengine/services/SaverService.java
@@ -2,13 +2,11 @@ package searchengine.services;
 
 import lombok.RequiredArgsConstructor;
 import org.jsoup.Connection;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import searchengine.model.PageEntity;
-import searchengine.model.SiteEntity;
+import searchengine.model.entities.PageEntity;
+import searchengine.model.entities.SiteEntity;
 import searchengine.repositories.PageRepository;
 import searchengine.repositories.SiteRepository;
 
@@ -26,13 +24,9 @@ public class SaverService {
         if (pageRepository.existsByPathAndSite(path, site)) {
             return null;
         }
-        return pageRepository.save(new PageEntity(site, path, response.statusCode(), response.body()));
-    }
-
-    @Retryable(maxAttempts = 14, backoff = @Backoff(delay = 200))
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public void updateSiteStatusTime(SiteEntity site) {
+        PageEntity page = pageRepository.save(new PageEntity(site, path, response.statusCode(), response.body()));
         site.setStatusTime(LocalDateTime.now());
         siteRepository.save(site);
+        return page;
     }
 }
